+++ server/indexTester.js
@@ -8,17 +8,15 @@ const authenticate = require("./middleware/authentication");
 
 const bodyParser = require("body-parser");
 const multer = require("multer");
-const fileUploader = require("express-fileupload");
-// configuring multer for file uploads
-// const storage = multer.memoryStorage();
-// const storage = multer.diskStorage({
-//   destination: './images',
-//   filename: (req, file, callBack) => {
-//     return callBack(null, `${file.filename}_${Date.now()}${path.extname(file.originalname)}`)
-//   }
-// });
-// const upload = multer({ storage: storage });
-const upload = multer();
+
+const upload = multer({
+  storage: multer.diskStorage({
+    destination: './images',
+    filename: (_req, file, callBack) => {
+      return callBack(null, `${Date.now()}_${file.originalname}`)
+    }
+  }),
+});
 
 const app = express();
 const port = process.env.SERVER_PORT;
@@ -27,19 +25,16 @@ app.use(cors());
 app.use(express.json());
 
 app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
 
-// const upload = multer({ storage: storage });
+// app.use("/api", authenticate);
 
 // apply multer to be used globally
-app.use(upload.none());
-app.use(fileUploader());
-
+app.use(upload.any());
 
 const jwtSecretKey = process.env.JWT_SECRET_KEY;
 
-app.get("/", (req, res) => {
-  res.send({ error: false, msg: "Server is Working Fine...."});
+app.get("/", (_req, res) => {
+  res.send({ error: false, msg: "Server is Working Fine...." });
 });
 
 app.post("/api/signup", async (req, res) => {
@@ -47,7 +42,7 @@ app.post("/api/signup", async (req, res) => {
     const { password, conPassword } = req.body;
     if (password !== conPassword) {
       console.log("password missmatch");
-      res.send({ error: true, message: "Password Mismatched"});
+      res.send({ error: true, message: "Password Mismatched" });
     } else {
       const hashPassword = await bcrypt.hash(req.body.password, 10);
       const user = { name: req.body.username, password: hashPassword, address: req.body.address };
@@ -62,11 +57,11 @@ app.post("/api/signup", async (req, res) => {
           user.password,
           user.address,
         ]);
-        
+
         const id = newUser.insertId;
-        const token = jwt.sign({ id }, jwtSecretKey, { expiresIn: "1h" })
-        
-        newUser.token = token;               
+        const token = jwt.sign({ id }, jwtSecretKey, { expiresIn: "1h" });
+
+        newUser.token = token;
         res.send({ error: false });
       }
     }
@@ -80,72 +75,73 @@ app.post("/api/login", async (req, res) => {
   try {
     const { username, password } = req.body;
     if (!username || !password) {
-      res.send({ error: true, message: "Username and Password Required!!"});
+      res.send({ error: true, message: "Username and Password Required!!" });
     } else {
       const [users] = await dbConnection.query("SELECT * FROM users WHERE name = ?", [username]);
       if (await bcrypt.compare(password, users[0].password)) {
         const id = users[0].id;
-        const token = jwt.sign({ id }, jwtSecretKey, { expiresIn: "10m" });
+        const token = jwt.sign({ id }, jwtSecretKey, { expiresIn: "1h" });
         console.log({ Longin: true, token });
-        res.set('authorization', `Bearer ${token}`).status(201).json({ token, error: false, name: users[0].name});
+        res.set("authorization", `Bearer ${token}`).status(201).json({ token, error: false, name: users[0].name });
       } else {
-        res.send({ error: true, message: "Wrong Pword"});
+        res.send({ error: true, message: "Wrong Pword" });
       }
     }
   } catch (error) {
     console.log("🚀 ~ file: index.js:59 ~ app.post ~ error:", error);
-    res.status(404).json({ error: true, message: "User does not exist"});
+    res.status(404).json({ error: true, message: "User does not exist" });
   }
 });
 
-app.get("/api/dashboard", authenticate, async (req, res) => {
+app.get("/api/dashboard", async (req, res) => {
   try {
     const query = "SELECT * FROM users WHERE id = ?";
 
     const [rows] = await dbConnection.query(query, [req.userId]);
-    if(rows.length === 0) {
-      res.status(404).json(({error: "User not found"}));
-    }else{
+    if (rows.length === 0) {
+      res.status(404).json({ error: "User not found" });
+    } else {
       const user = rows[0];
-      res.status(200).json({ name: user.name, address: user.address});
+      res.status(200).json({ name: user.name, address: user.address });
     }
   } catch (error) {
-    console.log("🚀 ~ file: index.js:113 ~ app.post ~ error:", error)    
+    console.log("🚀 ~ file: index.js:113 ~ app.post ~ error:", error);
   }
 });
 
-app.get("/api/products", authenticate, async (req, res) => {
+app.get("/api/products", async (_req, res) => {
   try {
     const query = "SELECT * FROM products";
 
     const [rows] = await dbConnection.query(query);
-    if(rows.length === 0) {
-      res.status(404).json(({error: "No products found"}));
-    }else{
+    if (rows.length === 0) {
+      res.status(404).json({ error: "No products found" });
+    } else {
       res.status(200).json(rows);
     }
   } catch (error) {
-    console.log("🚀 ~ file: index.js:113 ~ app.post ~ error:", error)
+    console.log("🚀 ~ file: index.js:113 ~ app.post ~ error:", error);
   }
 });
 
-// app.post("/api/add-products", authenticate, upload.single(''), async (req, res) => {
-app.post("/api/add-products", authenticate, async (req, res) => {
+// app.use("/api", authenticate);
+
+app.post("/api/add-products", async (req, res) => {
   try {
     const { name, size, price } = req.body;
-    const image = req.files.image;
-    if(!(name, size, price)) {
-      res.send({error: true, message: "Please input all fields."});
+    const image = req.files[0];
+    if (!(name, size, price)) {
+      res.send({ error: true, message: "Please input all fields." });
     } else {
       const query = "INSERT INTO products (name, size, price) VALUES (?, ?, ?);";
-      
+
       const [newProduct] = await dbConnection.query(query, [name, size, price]);
-      res.status(201).json({ error: false,  message: "Products Created", Products_name: newProduct.name });
-      console.log(image.originalname);
+      res.status(201).json({ error: false, message: "Products Created", Products_name: newProduct.name });
+      console.log('filename:', image.filename);
     }
   } catch (error) {
-    console.log("🚀 ~ file: index.js:113 ~ app.post ~ error:", error);
-    res.status(404).json({ error: true, message: "Failed to add products"});
+    console.log("🚀 ~ file: index.js:143 ~ app.post ~ error:", error);
+    res.status(404).json({ error: true, message: "Failed to add products" });
   }
 });
